src/lambda/requirements_lambda.txt
python-dotenv==1.0.1
unsloth==2024.12.12
unsloth_zoo==2024.12.7
xformers==0.0.27.post2
tf_keras==2.17.0
Jinja2==3.1.4
